Refactoring tests and CI
