+++ .github/workflows/ci.yml
@@ -207,3 +207,32 @@ jobs:
       if: env.CODECOV_TOKEN != null
       env:
         CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
+
+  doc_snippets:
+    name: Documentation Snippets
+    runs-on: ubuntu-latest
+    steps:
+    - uses: actions/checkout@v4
+    
+    - name: Set up Python
+      uses: actions/setup-python@v5
+      with:
+        python-version: '3.x'
+    
+    - name: Install Rust
+      uses: dtolnay/rust-toolchain@stable
+    
+    - name: Cache cargo registry
+      uses: actions/cache@v4
+      with:
+        path: |
+          ~/.cargo/registry
+          ~/.cargo/git
+          target
+        key: ${{ runner.os }}-cargo-doc-snippets-${{ hashFiles('**/Cargo.lock') }}
+        restore-keys: |
+          ${{ runner.os }}-cargo-doc-snippets-
+          ${{ runner.os }}-cargo-
+    
+    - name: Check documentation snippets
+      run: ./scripts/check-doc-snippets.sh
