+++ tests/filters/mod.rs
@@ -1,0 +1,19 @@
+//! Test module for nessus filters functionality
+
+#[cfg(test)]
+pub mod test_fixtures;
+
+#[cfg(test)]
+pub mod basic_operations;
+
+#[cfg(test)]
+pub mod logical_operations;
+
+#[cfg(test)]
+pub mod nested_operations;
+
+#[cfg(test)]
+pub mod operation_builder;
+
+#[cfg(test)]
+pub mod integration;
