+++ tests/projections/test_fixtures.rs
@@ -1,0 +1,26 @@
+//! Shared test fixtures for projection tests
+
+use nessus::FieldWitnesses;
+use serde::{Deserialize, Serialize};
+
+#[derive(Deserialize, Serialize, FieldWitnesses)]
+pub struct User {
+    pub id: String,
+    pub name: String,
+    pub email: String,
+    pub age: u32,
+}
+
+#[derive(Deserialize, Serialize, FieldWitnesses)]
+pub struct Address {
+    pub street: String,
+    pub city: String,
+    pub zip: String,
+}
+
+#[derive(Deserialize, Serialize, FieldWitnesses)]
+pub struct Profile {
+    pub bio: String,
+    pub location: String,
+    pub avatar_url: String,
+}
