+++ tests/updates/mod.rs
@@ -1,0 +1,8 @@
+//! Test module for nessus updates functionality
+
+pub mod basic_operations;
+pub mod builder_functionality;
+pub mod enum_types;
+pub mod integration;
+pub mod nested_operations;
+pub mod test_fixtures;
